// src/lib/api/generated/model/combinationRuleType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type CombinationRuleType = typeof CombinationRuleType[keyof typeof CombinationRuleType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CombinationRuleType = {
  enables: 'enables',
  excludes: 'excludes',
  discount: 'discount',
  free_with: 'free_with',
} as const;