// src/lib/api/generated/model/discountType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

/**
 * Tipo de descuento. Solo para ruleType=discount
 */
export type DiscountType = typeof DiscountType[keyof typeof DiscountType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;